# services/api/docs/swagger.yaml
components:
  schemas:
...
    contracts.DeploymentCreate:
      properties:
        deployed_by:
          type: string
        environment_id:
          type: string
        intent_digest:
          type: string
        release_id:
          type: string
        status:
          enum:
          - pending
          - rendered
          - pushed
          - reconciling
          - healthy
          - degraded
          - failed
          - rolled_back
          type: string
        status_reason:
          type: string
      required:
      - environment_id
      - release_id
      type: object
    contracts.DeploymentPageResult:
      properties:
        items:
          items:
            $ref: '#/components/schemas/contracts.DeploymentResponse'
          type: array
          uniqueItems: false
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
      type: object
    contracts.DeploymentResponse:
      properties:
        created_at:
          type: string
        deployed_at:
          type: string
        deployed_by:
          type: string
        environment_id:
          type: string
        id:
          type: string
        intent_digest:
          type: string
        release_id:
          type: string
        status:
          type: string
        status_reason:
          type: string
        updated_at:
          type: string
      type: object
    contracts.DeploymentUpdate:
      properties:
        deployed_at:
          type: string
        status:
          enum:
          - pending
          - rendered
          - pushed
          - reconciling
          - healthy
          - degraded
          - failed
          - rolled_back
          type: string
        status_reason:
          type: string
      type: object
...
    contracts.ErrorDetail:
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
      type: object
    contracts.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/contracts.ErrorDetail'
      type: object
    contracts.ModuleVersion:
      properties:
        name:
          type: string
        version:
          type: string
      type: object
...
    contracts.RenderJobCreate:
      properties:
        bundle_hash:
          type: string
        module_versions:
          items:
            $ref: '#/components/schemas/contracts.ModuleVersion'
          type: array
          uniqueItems: false
        renderer_version:
          type: string
      type: object
    contracts.RenderJobResponse:
      properties:
        bundle_hash:
          type: string
        deployment_id:
          type: string
        finished_at:
          type: string
        id:
          type: string
        module_versions:
          items:
            $ref: '#/components/schemas/contracts.ModuleVersion'
          type: array
          uniqueItems: false
        oci_digest:
          type: string
        oci_ref:
          type: string
        output_hash:
          type: string
        renderer_version:
          type: string
        signature_verified_at:
          type: string
        signed:
          type: boolean
        started_at:
          type: string
        status:
          type: string
        storage_uri:
          type: string
      type: object
    contracts.RenderJobUpdate:
      properties:
        bundle_hash:
          type: string
        finished_at:
          type: string
        module_versions:
          items:
            $ref: '#/components/schemas/contracts.ModuleVersion'
          type: array
          uniqueItems: false
        oci_digest:
          type: string
        oci_ref:
          type: string
        output_hash:
          type: string
        renderer_version:
          type: string
        signature_verified_at:
          type: string
        signed:
          type: boolean
        status:
          enum:
          - pending
          - running
          - succeeded
          - failed
          type: string
        storage_uri:
          type: string
      type: object
...
info:
  title: ""
  version: ""
openapi: 3.1.0
paths:
...
  /api/v1/deployments:
    get:
      description: List deployments with optional filtering and pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        schema:
          type: integer
      - description: 'Page size (default: 20)'
        in: query
        name: page_size
        schema:
          type: integer
      - description: Filter by release ID
        in: query
        name: release_id
        schema:
          type: string
      - description: Filter by environment ID
        in: query
        name: environment_id
        schema:
          type: string
      - description: Filter by status (pending, healthy, unhealthy, failed)
        in: query
        name: status
        schema:
          type: string
      - description: Filter by deployer
        in: query
        name: deployed_by
        schema:
          type: string
      - description: Filter by deployment date (RFC3339)
        in: query
        name: since
        schema:
          type: string
      - description: Filter by deployment date (RFC3339)
        in: query
        name: until
        schema:
          type: string
      - description: Sort field (created_at, deployed_at)
        in: query
        name: sort_by
        schema:
          type: string
      - description: Sort order (asc, desc)
        in: query
        name: sort_order
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.DeploymentPageResult'
          description: Paginated list of deployments
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: List deployments
      tags:
      - deployments
    post:
      description: Create a new deployment for a release to an environment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contracts.DeploymentCreate'
        description: Deployment creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.DeploymentResponse'
          description: Created deployment
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Release or environment not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Create a new deployment
      tags:
      - deployments
  /api/v1/deployments/{deployment_id}/render-job:
    get:
      description: Retrieve the render job associated with a deployment
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: deployment_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.RenderJobResponse'
          description: Render job details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid deployment ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Render job not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Get render job for a deployment
      tags:
      - deployments
    patch:
      description: Update the render job associated with a deployment
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: deployment_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contracts.RenderJobUpdate'
        description: Render job update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.RenderJobResponse'
          description: Updated render job
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Render job not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Update a render job
      tags:
      - deployments
    post:
      description: Create a new render job for an existing deployment
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: deployment_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contracts.RenderJobCreate'
        description: Render job creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.RenderJobResponse'
          description: Created render job
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Deployment not found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Render job already exists for this deployment
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Create a render job for a deployment
      tags:
      - deployments
  /api/v1/deployments/{id}:
    delete:
      description: Delete a deployment
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: Deployment deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid deployment ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Deployment not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Delete a deployment
      tags:
      - deployments
    get:
      description: Retrieve a single deployment by its ID
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.DeploymentResponse'
          description: Deployment details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid deployment ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Deployment not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Get a deployment by ID
      tags:
      - deployments
    patch:
      description: Update a deployment's status and status reason
      parameters:
      - description: Deployment ID (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contracts.DeploymentUpdate'
        description: Deployment update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.DeploymentResponse'
          description: Updated deployment
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Deployment not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Invalid status transition
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contracts.ErrorResponse'
          description: Internal server error
      summary: Update a deployment
      tags:
      - deployments
...